requirements.txt
# Testing
pytest>=7.0.0
pytest-playwright>=0.4.0

# Validation
pydantic>=2.0.0

# Test data generation
faker>=20.0.0

# Logging
loguru>=0.7.0
